# Optimize handle_interrupt(Exception => ..) as a common case