@ioquatix 2023-08-25
This looks pretty nice to me.